+++ VE_Conductor/Framework/TexTable.cpp
@@ -86,7 +86,7 @@ void TexTable::SetNumofCols(int num)
 void TexTable::SetColTitles(const std::vector<wxString>& titles)
 {
    for ( unsigned int i=0; 
-         (i < titles.size() ) && ( i < GetNumberCols() ); ++i)
+         (i < (size_t) titles.size() ) && ( i < (size_t) GetNumberCols() ); ++i)
    {
       SetColLabelValue( i, titles[i] );
    }
@@ -96,9 +96,9 @@ void TexTable::SetColAlignments(const std::vector<int>& alignments)
 {
    m_align = alignments;
    for ( unsigned int i=0; 
-         (i < alignments.size() ) && ( i < GetNumberCols() ); ++i )
+         (i < (size_t) alignments.size() ) && ( i < (size_t) GetNumberCols() ); ++i )
    {
-      for ( unsigned int j = 0; j < GetNumberRows(); ++j)
+      for ( unsigned int j = 0; j < (size_t) GetNumberRows(); ++j)
       {
          SetCellAlignment( j, i, alignments[i], wxALIGN_CENTER);
       }
